+++ source/CHANGELOG.md
@@ -8,6 +8,7 @@ CHANGED:
 * Analysis function GoodnessOfFit: slight improvement to scaling of volume error for display in radar chart #98
 
 FIXED:
+* Detect empty time series before attempting to import them
 * Some fixes to reading and writing HYSTEM-EXTRAN REG format #106
 
 Version 2.4.0